+++ src/com/jrsoft/fri/dagl/entity/DaxxGateway.java
@@ -9,7 +9,7 @@ public class DaxxGateway implements java.io.Serializable {
 
 	// Fields
 
-	private int id;				//
+	private Long id;				//
 	private String type;				//电梯网关类型
 	private String flow;				//取流方式
 	private String net;				//入网方式
@@ -50,11 +50,11 @@ public class DaxxGateway implements java.io.Serializable {
 
 	// Property accessors
 
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
